backend/requirements.txt
Flask==2.1.1
python-jose==3.3.0
boto3==1.20.24
pytest==6.2.5
pytest-mock==3.6.1
python-dotenv
botocore
pyjwt
dnspython
pymongo[srv]
bcrypt